+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 ### Electronic horizon
 
 * Extended `RoadGraph.Edge.Metadata` properties by adding `freeFlowSpeed`, `constrainedFlowSpeed`, `isFerry`, `isRoundabout`, `countryCodeISO3`, `surface` and `sapaType` members.
+* Added `ElectronicHorizonConfig.enableEnhancedDataAlongEH` option to allow requesting additional data from ADAS layer.
+* Added `RoadGraph.adasAttributes(edgeIdentifier:)` to fetch `ADASAttributes` on a given edge.
 
 ### Routing